genesis_rl/pyproject.toml
[project]
name = "genesis-rl"
version = "0.1.0"
description = "SO101 Pick-and-Place RL training with Genesis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "genesis-world>=0.4.6",
    "gymnasium>=1.3.0",
    "ipython>=9.13.0",
    "numpy>=2.4.4",
    "pyyaml>=6.0.3",
    "rsl-rl-lib>=5.2.0",
    "tensorboard>=2.20.0",
]
